requirements.txt
# Core Data Transformation
dbt-snowflake==1.7.1
dbt-core==1.7.4

# Data Generation
faker==22.6.0
numpy==1.26.3
pandas==2.2.0

# Streamlit Dashboard
streamlit==1.31.0
plotly==5.18.0
snowflake-connector-python==3.7.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
